pyproject.toml
[tool.poetry]
name = "tripkro"
version = "0.1.0"
description = ""
authors = ["shehroz javed <shehroz.javed@synar.es>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
djangorestframework-simplejwt = "^5.3.1"
psycopg = "^3.1.18"
pre-commit = "^3.6.2"
drf-standardized-errors = "^0.13.0"
django-debug-toolbar = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
